refactor: refactor article image styles, prop types and coverage to 100%
