+++ packages/article-image/src/article-image-prop-types.js
@@ -1,18 +1,18 @@
 import PropTypes from "prop-types";
 
-export const articleImagePropTypes = {
+export const propTypes = {
+  captionOptions: PropTypes.shape({
+    caption: PropTypes.string,
+    credits: PropTypes.string
+  }),
   imageOptions: PropTypes.shape({
     display: PropTypes.oneOf(["primary", "secondary", "inline"]),
     ratio: PropTypes.string,
     url: PropTypes.string.isRequired
-  }).isRequired,
-  captionOptions: PropTypes.shape({
-    caption: PropTypes.string,
-    credits: PropTypes.string
-  })
+  }).isRequired
 };
 
-export const articleImageDefaultPropTypes = {
+export const defaultPropTypes = {
   captionOptions: {
     caption: null,
     credits: null
