+++ packages/article-image/src/article-image.base.js
@@ -1,39 +1,25 @@
 import React from "react";
-import { View, StyleSheet } from "react-native";
-import { ModalImage } from "@times-components/image";
+import { View } from "react-native";
 import Caption from "@times-components/caption";
-import { spacing } from "@times-components/styleguide";
+import { ModalImage } from "@times-components/image";
+import styles from "../styles";
 import InsetCaption from "./inset-caption";
 import InlineImage from "./inline-image";
-
-import {
-  articleImagePropTypes,
-  articleImageDefaultPropTypes
-} from "./article-image-prop-types";
-
-const styles = StyleSheet.create({
-  secondaryImage: {
-    width: "50%"
-  },
-  secondaryCaption: {
-    paddingLeft: spacing(2),
-    paddingTop: 0,
-    width: "50%"
-  }
-});
-
-const captionStyle = {
-  secondary: {
-    container: {
-      paddingTop: 0
-    }
-  }
-};
+import { propTypes, defaultPropTypes } from "./article-image-prop-types";
 
 const renderCaption = (display, caption, credits) => {
   if (!caption && !credits) {
     return null;
   }
+
+  const captionStyle = {
+    secondary: {
+      container: {
+        paddingTop: 0
+      }
+    }
+  };
+
   const CaptionComponent =
     display === "primary" ? (
       <InsetCaption caption={caption} credits={credits} />
@@ -78,7 +64,7 @@ const ArticleImage = ({ imageOptions, captionOptions }) => {
   ];
 };
 
-ArticleImage.propTypes = articleImagePropTypes;
-ArticleImage.defaultProps = articleImageDefaultPropTypes;
+ArticleImage.propTypes = propTypes;
+ArticleImage.defaultProps = defaultPropTypes;
 
 export default ArticleImage;
